requirements.txt
# Core dependencies
asyncio-mqtt>=0.16.1
aiofiles>=23.2.1
python-dotenv>=1.0.0

# Machine Learning
torch>=2.0.0
torchvision>=0.15.0
gymnasium>=0.29.0
stable-baselines3>=2.1.0
optuna>=3.4.0
joblib>=1.3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Natural Language Processing
nltk>=3.8.1

# Quantum-resistant cryptography
cryptography>=41.0.0
scipy>=1.11.0

# Cloud integrations
boto3>=1.34.0
botocore>=1.34.0
azure-identity>=1.15.0
azure-mgmt-compute>=30.3.0
azure-mgmt-network>=30.0.0
azure-mgmt-security>=5.0.0
azure-mgmt-monitor>=5.0.0
google-cloud-compute>=1.14.0
google-cloud-securitycenter>=1.14.0
google-cloud-monitoring>=2.16.0
google-auth>=2.23.0
kubernetes>=28.1.0
docker>=6.1.0
openstacksdk>=1.3.0

# Voice and AI
SpeechRecognition>=3.10.0
pyttsx3>=2.90
openai>=1.3.0

# Networking and protocols
requests>=2.31.0
websockets>=11.0.0
paramiko>=3.3.0

# Database and storage
redis>=5.0.0
psycopg2-binary>=2.9.0
PyMySQL>=1.1.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Monitoring and logging
prometheus-client>=0.17.0
structlog>=23.1.0

# Security
PyJWT>=2.8.0
passlib>=1.7.4
bcrypt>=4.0.0
